can archive/unarchive notes

+++ 3rd term/Lab2/main_window.cpp
@@ -1,8 +1,11 @@
 #include "main_window.h"
 #include "ui_main_window.h"
 
-QString DATA_PATH = "data/notesData.txt";
-QString TEXT_PATH = "data/notesText/";
+QString MAIN_DATA_PATH = "data/main_notesData.txt";
+QString MAIN_TEXT_PATH = "data/main_notesText/";
+
+QString ARCHIVE_DATA_PATH = "data/archive_notesData.txt";
+QString ARCHIVE_TEXT_PATH = "data/archive_notesText/";
 
 Note* Note::currentNote = nullptr;
 
@@ -10,17 +13,21 @@ MainWindow::MainWindow(QWidget *parent)
     : QMainWindow(parent)
     , ui(new Ui::MainWindow)
     , main_model(new QStandardItemModel)
+    , archive_model(new QStandardItemModel)
 {
     ui->setupUi(this);
 
     ui->listView->setModel(main_model);
-    readNotesFromFile(DATA_PATH, main_model);
+    readNotesFromFile(MAIN_DATA_PATH, MAIN_TEXT_PATH, main_model);
+    readNotesFromFile(ARCHIVE_DATA_PATH, ARCHIVE_TEXT_PATH, archive_model);
 }
 
 MainWindow::~MainWindow(){
     //saving any changes
-    writeNotesDataToFile(DATA_PATH, notes);
-    writeNotesTextToFiles();
+    writeNotesDataToFile(MAIN_DATA_PATH, main_notes);
+    writeNotesDataToFile(ARCHIVE_DATA_PATH, archived_notes);
+    writeNotesTextToFiles(MAIN_TEXT_PATH, main_notes);
+    writeNotesTextToFiles(ARCHIVE_TEXT_PATH, archived_notes);
     delete ui;
 }
 
@@ -33,8 +40,8 @@ void MainWindow::on_add_note_clicked()
 
     if(Note::currentNote != nullptr){
         //if "Save" was clicked add note to list
-         notes.push_back(Note::currentNote);
-         main_model->insertRow(notes.size()-1, new QStandardItem(Note::currentNote->title));
+         main_notes.push_back(Note::currentNote);
+         main_model->insertRow(main_notes.size()-1, new QStandardItem(Note::currentNote->title));
          Note::currentNote = nullptr;
     }
 
@@ -47,8 +54,10 @@ void MainWindow::on_listView_clicked(const QModelIndex &index)
     ui->archive_btn->setEnabled(true);
     ui->delete_btn->setEnabled(true);
 
-    Note::currentNote = notes.at(index.row());
-    qDebug()<<Note::currentNote->title;
+    if(ui->listView->model() == main_model){
+        Note::currentNote = main_notes.at(index.row());
+    } else
+        Note::currentNote = archived_notes.at(index.row());
 }
 
 void MainWindow::on_open_btn_clicked()
@@ -62,33 +71,77 @@ void MainWindow::on_open_btn_clicked()
 
         //remove old text
         int index = ui->listView->currentIndex().row();
-        QString old_text_file_path = TEXT_PATH+Note::dateToFileName(notes[index]->editTime);
+        QString old_text_file_path;
+
+        if(ui->listView->model() == main_model){
+            old_text_file_path = MAIN_TEXT_PATH+Note::dateToFileName(main_notes[index]->editTime);
+        } else
+            old_text_file_path = ARCHIVE_TEXT_PATH+Note::dateToFileName(archived_notes[index]->editTime);
+
         QFile old_text_file(old_text_file_path);
         old_text_file.remove();
 
-        notes.erase(notes.begin()+index);
-        notes.insert(notes.begin()+index, Note::currentNote);
+        if(ui->listView->model() == main_model){
+            main_notes.erase(main_notes.begin()+index);
+            main_notes.insert(main_notes.begin()+index, Note::currentNote);
+        } else {
+            archived_notes.erase(archived_notes.begin()+index);
+            archived_notes.insert(archived_notes.begin()+index, Note::currentNote);
+        }
+    }
+}
+
+void MainWindow::on_archive_btn_clicked()
+{
+    int index = ui->listView->currentIndex().row();
+
+    if(ui->listView->model() == main_model){
+        QFile file(MAIN_TEXT_PATH+Note::dateToFileName(main_notes[index]->editTime));
+        file.remove();
+
+        archived_notes.push_back(main_notes[index]);
+        main_notes.erase(main_notes.begin()+index);
+
+        archive_model->insertRow(archived_notes.size()-1, new QStandardItem(archived_notes.last()->title));
+        main_model->removeRow(index);
+    }
+}
+
+void MainWindow::on_delete_btn_clicked()
+{
+    int index = ui->listView->currentIndex().row();
+
+    if(ui->listView->model() == main_model){
+        QFile file(MAIN_TEXT_PATH+Note::dateToFileName(main_notes[index]->editTime));
+        file.remove();
+        main_notes.erase(main_notes.begin()+index);
+        main_model->removeRow(index);
+    }else {
+        QFile file(ARCHIVE_TEXT_PATH+Note::dateToFileName(archived_notes[index]->editTime));
+        file.remove();
+        archived_notes.erase(archived_notes.begin()+index);
+        archive_model->removeRow(index);
     }
 
 }
 
 
-void MainWindow::on_delete_btn_clicked()
+void MainWindow::on_main_list_btn_clicked()
 {
-    int index = ui->listView->currentIndex().row();
-    QFile file(TEXT_PATH+Note::dateToFileName(notes[index]->editTime));
-    file.remove();
-    notes.erase(notes.begin()+index);
-    main_model->removeRow(index);
+    ui->listView->setModel(main_model);
 }
 
+void MainWindow::on_archive_list_btn_clicked()
+{
+    ui->listView->setModel(archive_model);
+}
 
 //----------------------------------------------------------------------FILES------------------------------------------------------------------------
 
 
-void MainWindow::readNotesFromFile(QString fileName, QStandardItemModel *model)
+void MainWindow::readNotesFromFile(QString dataFileName, QString textFolderName, QStandardItemModel *model)
 {
-    QFile dataNotes(fileName);
+    QFile dataNotes(dataFileName);
     if (!dataNotes.open(QIODevice::ReadOnly | QIODevice::Text))
         return;
 
@@ -112,7 +165,7 @@ void MainWindow::readNotesFromFile(QString fileName, QStandardItemModel *model)
 
         //read text
         QString fileName = Note::dateToFileName(createTime);
-        QFile textNotes(TEXT_PATH+fileName);
+        QFile textNotes(textFolderName+fileName);
         if (!textNotes.open(QIODevice::ReadOnly | QIODevice::Text))
             return;
         QString text = textNotes.readAll();
@@ -129,21 +182,28 @@ void MainWindow::readNotesFromFile(QString fileName, QStandardItemModel *model)
 
 void MainWindow::addNoteToList(Note* note, QStandardItemModel *model)
 {
-    int position = getNotePositionInFile(0, std::max(0,notes.size()-1), *note);
-    notes.insert(notes.begin() + position, note);
+    int position;
+    if(model == main_model){
+        position = getNotePositionInFile(0, std::max(0,main_notes.size()-1), *note, main_notes);
+        main_notes.insert(main_notes.begin() + position, note);
+    } else {
+        position = getNotePositionInFile(0, std::max(0,archived_notes.size()-1), *note, archived_notes);
+        archived_notes.insert(archived_notes.begin() + position, note);
+    }
     model->insertRow(position, new QStandardItem(note->title));
 }
 
-int MainWindow::getNotePositionInFile(int left, int right, Note& note)
+int MainWindow::getNotePositionInFile(int left, int right, Note& note, QVector<Note*>& list)
 {
     if (left == right)
         return left;
 
     int middle = (left+right)/2;
-    if (notes[middle]->editTime > note.editTime)
-        getNotePositionInFile(left, middle, note);
+
+    if (list[middle]->editTime > note.editTime)
+        getNotePositionInFile(left, middle, note, list);
     else
-        getNotePositionInFile(middle+1, right, note);
+        getNotePositionInFile(middle+1, right, note, list);
 
     return left;
 }
@@ -170,20 +230,22 @@ void MainWindow::writeNotesDataToFile(QString fileName, QVector<Note*>& list)
     dataNotes.close();
 }
 
-void MainWindow::writeNotesTextToFiles()
+void MainWindow::writeNotesTextToFiles(QString folderName, QVector<Note*>& list)
 {
-    for (int i = 0; i < notes.size(); i++)
+    for (int i = 0; i < list.size(); i++)
     {
-        QString fileName = Note::dateToFileName(notes[i]->editTime);
-        QFile textNotes(TEXT_PATH+fileName);
+        QString fileName = Note::dateToFileName(list[i]->editTime);
+        QFile textNotes(folderName+fileName);
         if (!textNotes.open(QIODevice::WriteOnly | QIODevice::Text))
             return;
 
         QTextStream out(&textNotes);
 
-        out << notes[i]->text;
+        out << list[i]->text;
 
         out.flush();
         textNotes.close();
     }
 }
+
+
